python_env/requirements.txt
request
python_dotenv